fix(ci): clippy and two intra-doc links rustdoc reads as broken
clippy:
  - hc_mult is already usize; the cast was a no-op.
  - verify_d takes tokens as &[u32] already, unlike the fixed-width arrays
    the K=2/3/4 paths take, so &tokens[..] reslices the whole range.

rustdoc:
  - [`Qwen4ExpMtpModule`] is not in scope in that module, so the brackets
    were a broken link rather than a link.
  - [`hc_post`] likewise: the items are hc_post_site / hc_post_lowrank.
  - `scores[b]` had its [b] read as a link target. Both predate this
    stack and fail the docs job on any PR that touches the crate; same
    class as 924864cc.

+++ crates/spark-model/src/layers/qwen3_ssm/trait_decode_batched_hc.rs
@@ -147,7 +147,7 @@ impl Qwen3SsmLayer {
                         "hc batched verify: PLE has {} host id(s) for {num_tokens} row(s)",
                         host.len()
                     );
-                    let row_stride = (hc.hc_mult as usize) * h * 4; // highway rows are FP32
+                    let row_stride = hc.hc_mult * h * 4; // highway rows are FP32
                     let mut row0 = 0usize;
                     for (i, state) in states.iter_mut().enumerate() {
                         let ki = *ks.get(i).ok_or_else(|| {
